overflow in expression corrected: however this function is not longer used ...

+++ src/Varie/FunzRand.cpp
@@ -127,7 +127,7 @@ double unifRand()
 //    }
     
     //   void seed();
-    return rand() / double(RAND_MAX+1);
+    return rand() / (RAND_MAX+1.0);
 //    return rand_r(&seed_r) / double(RAND_MAX);
     //    return 2.0*rand()/RAND_MAX - 1;
